Added basic decompilation support.
Supports only labels and text right now, but at least it does something.
Basic infrastructure is in place, so now all that needs to be done is add
support for more node types.

Also need to figure out what's up with the children-less label nodes and if
I need to indent stuff.