Add a "--comparable" switch
Make it easier to compare AST dumps of different rpyc's by using the new
"--comparable" switch. Currently, its only effect is to suppress properties
such as file names and modification times from the dumps. In the future, I
plan to add other features to it, such as triggering the insertion of blank
lines to make line numbers match.