Handle ATL's loc being ('', 0)
If a statement ends with a colon but doesn't have a block after it, the ATL's
loc property gets set to ('', 0). Since that should be a syntax error, and might
be in a future Ren'Py release, and all that it affects is loc, we'll only
replicate it in comparable mode.