Rework comparable
The --comparable option is now dump-only. The new option --line-numbers
makes line numbers match when decompiling, and makes them be compared when
dumping with --comparable. The --no-file-metadata option is gone, since a
combination of the first two can replace it. Also, some minor fixes like
support for PyExpr with comparable, and a typo in the banner.