Be smarter about where to put "as" and "tag"
Instead of always putting "as" and "tag" at the beginning of a screen block,
put them either (in order of preference) on the same line as some other
keyword, on a blank line before a child, or at the end of the screen block.
This will fix some cases of line numbers not matching.